add posters

+++ app/src/main/java/com/egaragul/androidfundametals/ui/movies/data/Genre.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Genre(
-    val id: Int, val name: String
+    val id: Long, val name: String
 )
